add namespace field to pulumi.json
Add a namespace field to the `pulumi.json` file.  This will be used to
disambiguate package names for components, any TF provider and anyone
else that wants to use it.  The idea is to use this to generate
packages that can be imported as `@example/xyz` or `example_xyz`
instead of `@pulumi/xyz` or `pulumi_xyz`, when they are really not
coming from pulumi.

Pulumiverse packages will also be able to use this.

The codegen implementations for each of the languages will follow in
separate PRs.